Split access token and authorization code models
